// src/types/RocketSpecificInfo.tsx
export interface RocketPayload {
  name: string;
  type: string;
  active: boolean;
  stages: number;
  boosters: number;
  cost_per_launch: number;
  success_rate_pct: number;
  first_flight: string;
  country: string;
  company: string;
  wikipedia: string;
  description: string;
  id: string;
}

export interface RocketSpecificInfo {
  status: string;
  payload: RocketPayload;
}